git-svn-id: http://svn.seqan.de/seqan/trunk@1724 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/graph/graph_interface.h
@@ -95,7 +95,7 @@ struct Spec<Graph<TSpec> const>
 template<typename TSpec>
 struct Size<Graph<TSpec> > 
 {
-	typedef size_t Type;
+	typedef unsigned int Type;
 };
 
 //////////////////////////////////////////////////////////////////////////////
@@ -103,7 +103,7 @@ struct Size<Graph<TSpec> >
 template<typename TSpec>
 struct Size<Graph<TSpec> const>
 {
-	typedef size_t Type;
+	typedef unsigned int Type;
 };
 
 //////////////////////////////////////////////////////////////////////////////
